Only the board can now upload sounds for users
