+++ resources/views/users/dashboard/dashboard.blade.php
@@ -18,8 +18,6 @@
 
             @include('users.dashboard.includes.cardinfo')
 
-            @include('users.dashboard.includes.omnomcomsound')
-
         </div>
 
         <div class="col-xl-3 col-md-12">
